restructuring for new components AuthButton and LaguageSwitcher and useQueryParams moved to hooks from utils folder
