+++ src/components/AuthButton/AuthButton.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../../hooks/useAuth';
 
 export const AuthButton = () => {
   const { user, loading, handleSignIn, handleSignOut } = useAuth();
